Add no_std compatibility
By adding a default feature of "std" this is a backwards compatible
change that allows highway-rs to be built without the standard library
if one disables all features.

No APIs are changed in the process (other than std APIs like Debug or
Write). There is a behavior change when determining the best
implementation: since CPU detection is not available to `no_std` the
builder will always choose the portal implementation, so users who want
to use SSE 4.2 or AVX 2 will need to explicitly construct them with
`force_new` and the associated risk with calling them.