Make WasmHash::new infallible
The mere presence of Wasm SIMD instructions will cause older Wasm
compilers to fail even if the Wasm SIMD instruction is never executed.
With that huge caveat, feature detection inside Wasm will not be until
an inordinate amount of time has passed, as first feature detection
needs to be implemented and then we must wait until the runtimes without
feature detections fall out of use.

Since I do not want to give the impression that feature detection is
available with the `Option` return on `WasmHash::new`, it has been
changed to always be infallible.