Add aarch64 neon support
Based on discussions here:
https://github.com/nickbabcock/highway-rs/pull/51#discussion_r815247129

The following aren't stable:
- std::is_aarch64_feature_detected
- aarch64 target features

We're not really able to detect at runtime or compile time if neon
support is available

The good news is that it seems reasonable to assume the aarch64
environment is neon capable. If a use case is found where neon is not
available, we can patch that in later.

Benchmarks on a t4g.small

| payload (bytes) by MiB/s | 0.7 | 0.8  | improvement |
|--------------------------|-----|------|-------------|
| 1                        | 4.4 | 12.2 | 2.8x        |
| 65536                    | 930 | 3897 | 4.2x        |