Complete rename HighwayBuilder -> HighwayHasher
Back in v0.6.0, the `HighwayHasher` alias was introduced for
`HighwayBuilder`. I was never happy with the `HighwayBuilder` moniker as
it doesn't feel right for the main entrypoint of an API to use `Builder`
when a better name exists.

Thus the `HighwayHasher` alias was born as a way to test the waters
without needless backwars incompatibilities.

And now after a year and a half, it seems right to fully commit to the
alias