Eliminate bound checks
Looking at the release assembly, I found instances where the compiler
couldn't tell that the code couldn't panic, so I massaged it so the
compiler could better infer.

This should mean that the hashers (or at least the portable hasher) is
panic-less.

Throughput improved by about a few percentage points. Nothing exciting,
and I was able to verify that the amount of assembly decreased.