Add aarch64 testing terraform config
When testing the linux aarch64 code on a real platform (AWS graviton),
it is a bit of a hassle walking through all the steps to instantiate,
configure, and then run the test suite. So I created a terraform
configuration that does it all for me (a la infrastructure as code).

Right now, the config is a too tailored to me (assumes the default
network policy allows inbound ssh and the ssh key is named "me"). But,
I'm a terraform noob, so I welcome implementation suggestions.