Allow checkpointing and restoration of hash state
If one is hashing a stream of data over a long period of time, it
becomes conduscive to be able to checkpoint the hash state to allow one
to recover the state without rehashing the rest of the data.

It also allows great flexibility on how one wants to hash data.

I'm not tied to this exact API. Hard coding the number of bytes is
potentially brittle, but it does remove any chance of a fallible write
or read during serialization / deserialization. I don't see format
changing for some time.