Add public inherent methods to hashing implementation
When someone uses a highway hasher without importing the HighwayHash
trait, they get a cryptic "private method" error because trait methods
aren't available without the trait import and it happens to be that the
backing implementation has the same function name.

Rust-analyzer shows unhelpful error messages and no quick fixes.

This seems less than ideal, so this commit makes implementations public
so that the trait import is not needed.